requirements-api.txt
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=22.0.0
google-cloud-bigquery>=3.20.0
scikit-learn>=1.4.0
